Make tool page internal links focus buttons
We had a complaint that the internal links remained
in their 'hover state' even though the link had been
clicked — the user would discover this when scrolling
back up. This was due to the focus state remaining on
the button (as expected).

It's probably a good idea to move the focus state
to the part of the page the user has chosen to visit.
This has to be balanced with the expectations of
non-disabled clients that there not be blue borders
around everything.

We've chosen to focus on the tools page issue and
change the focus to the first button in each tool
section when it is navigated to with the internal
nav.

The accessibility issue (internal scroll without
focus change) remains on other pages (e.g. the
home page, but I'm not sure how we might solve it
while meeting client expectations.