Don't show 'Open tool' button if no URL present
Previously we would show tools without an active
website in 'coming soon' — but now apparently there
are tools which are complete but for whatever
reason have no URL. So we let tools only have
one button.

This was surprisingly tricky to do cleanly. I
might just be out of sleep or something.