Make body copy visited link same colour as link
This is a bit of a hassle because there are plenty
of un-class-ified links in the site that would
be affected by a plain `a:visited` selector (and
pseudoselectors have a relatively high precedence).

This is a fairly common issue in CMS-driven sites
as you don't usually have classes on those links.
One good solution is to have a class called
`.body-copy` or `.formatted-text` which applies
all that copy with child selectors. In this case
we get a similar effect by targeting links in
paragraph tags, since they only usually appear in
formatted text.

We'll need to go through the whole site and look
for mistaken applications though. I guess this is
a good argument for normalize.css!