Refine events filtering front-end transitions
Notably here, the content would sometimes appear
while the fading transition was in progress, which
looked a little messy. We fix this by waiting until
the fadeOut is complete before we request any data.

Ideally we'd do this with a promise so we can make
the request before we start the animation but only
change the content after both the fadeOut and the
request are finished. But I'm keen to avoid changing
too much code here where it's unwarranted without
good reason.

Conversely, if the fadeOut completed before the request
the page would suddenly collapse awkwardly and show
the footer until the content came back. The above
change exposed this, so we've fixed that too by
changing the `fadeOut` calls to `animate` calls
that just change the opacity without setting
`display: none`. This means our content is still
there to extend the page until we have new content
to shove in there.

Also a few minor changes to the surrounding code
(e.g. removing globals, extraneous whitespace, etc)