Order events by date descending
Misread the requirements again to be clearer than
they actually were. Client asked for soonest events
first, but then specified that an event that had passed
should be first. Needs clarification so we remove
the filtering and just order by date desc.

Further to this, the sorting by date wasn't actually
working anyway, as Wordpress sorts dates as strings
by default (unhelpfully) and we need to manually
tell it to not do that. Just happened that it looked
right with our set of data locally...